matplotlib 16

+++ sensi-telegram/auth.py
@@ -139,17 +139,22 @@ class Auth:
             msgGraphics ="Olá *" + self.name +"*"+ self.graphic.getInfo() + "dos últimos 3 dias. Lembrando que os gráficos devem demorar alguns minutinhos para chegar :)"
             bot.send_message(self.chat_id, msgGraphics, parse_mode="markdown")
 
-            self.graphic.makeGraphicAll(700)
+            regs = self.graphic.makeGraphicAll(700)
 
-            for tag in query:
-                msgTag = "SensiTag: " + tag[2] + "\nMAC: " + tag[1]
-                bot.send_message(self.chat_id, msgTag, parse_mode="markdown")
-                bot.send_photo(self.chat_id, open(tempDir + str(tag[1]) + "_Temperatura.png", "rb"))
-                os.system("rm " + tempDir + str(tag[1]) + "_Temperatura.png")
-                bot.send_photo(self.chat_id, open(tempDir + str(tag[1]) + "_Umidade.png", "rb"))
-                os.system("rm " + tempDir + str(tag[1]) + "_Umidade.png")
-                bot.send_photo(self.chat_id, open(tempDir + str(tag[1]) + "_Bateria.png", "rb"))
-                os.system("rm " + tempDir + str(tag[1]) + "_Bateria.png")
+            if regs == 1:
+                
+                for tag in query:
+                    msgTag = "SensiTag: " + tag[2] + "\nMAC: " + tag[1]
+                    bot.send_message(self.chat_id, msgTag, parse_mode="markdown")
+                    bot.send_photo(self.chat_id, open(tempDir + str(tag[1]) + "_Temperatura.png", "rb"))
+                    os.system("rm " + tempDir + str(tag[1]) + "_Temperatura.png")
+                    bot.send_photo(self.chat_id, open(tempDir + str(tag[1]) + "_Umidade.png", "rb"))
+                    os.system("rm " + tempDir + str(tag[1]) + "_Umidade.png")
+                    bot.send_photo(self.chat_id, open(tempDir + str(tag[1]) + "_Bateria.png", "rb"))
+                    os.system("rm " + tempDir + str(tag[1]) + "_Bateria.png")
+            else:
+                bot.send_message(self.chat_id, "Sensi aqui!\nParece que um de suas SensiTags não possui registros.", parse_mode="markdown")
+
         else:
             bot.send_message(self.chat_id, "Sensi aqui!\nParece que o seu sistema não possui SensiTags Cadastradas.", parse_mode="markdown")
